backend/requirements.txt
argcomplete==2.0.0
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
boto3==1.21.0
botocore==1.24.0
certifi==2021.10.8
cffi==1.15.0
cfn-flip==1.3.0
charset-normalizer==2.0.12
click==8.0.3
cryptography==36.0.1
Deprecated==1.2.13
Django==4.0.2
django-cors-headers==3.11.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
django-filter==21.1
django-mysql==4.5.0
django-nine==0.2.5
django-oauth-toolkit==1.7.0
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
durationpy==0.5
elastic-transport==8.1.0
elasticsearch==7.17.1
elasticsearch-dsl==7.4.0
future==0.18.2
h11==0.13.0
hjson==3.0.2
idna==3.3
inflection==0.5.1
jmespath==0.10.0
jwcrypto==1.0
kappa==0.6.0
lxml==4.8.0
oauthlib==3.2.0
opensearch-py==1.0.0
outcome==1.1.0
placebo==0.9.0
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
python-cas==1.6.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-aws4auth==1.1.1
s3transfer==0.5.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
troposphere==3.2.2
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
wsgi-request-logger==0.4.6
wsproto==1.0.0
zappa==0.54.1
